pyproject.toml
[project]
name = "roomba-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "fastapi-mcp>=0.3.3",
    "paho-mqtt>=2.1.0",
    "uvicorn>=0.34.2",
]
